requirements.txt
astroid==2.5
attrs==21.4.0
isort==4.3.21
lazy-object-proxy==1.7.1
marshmallow==3.6.1
marshmallow-sqlalchemy==0.23.1
mccabe==0.6.1
more-itertools==8.12.0
packaging==21.3
pluggy==0.13.1
py==1.11.0
pylint==2.5.3
pyparsing==3.0.8
pytest==5.4.1
SQLAlchemy==1.3.18
toml==0.10.2
wcwidth==0.2.5
wrapt==1.12.1
python-dateutil==2.8.2